api/requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=api/requirements/base.txt api/requirements/base.in
#
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
cent==4.1.0
    # via -r api/requirements/base.in
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via uvicorn
cryptography==37.0.4
    # via python-jose
dateparser==1.1.1
    # via -r api/requirements/base.in
ecdsa==0.18.0
    # via python-jose
fastapi==0.79.0
    # via -r api/requirements/base.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r api/requirements/base.in
idna==3.3
    # via
    #   anyio
    #   requests
    #   rfc3986
importlib-metadata==4.12.0
    # via markdown
markdown==3.4.1
    # via -r api/requirements/base.in
orjson==3.7.11
    # via -r api/requirements/base.in
passlib==1.7.4
    # via -r api/requirements/base.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.9.1
    # via
    #   -r api/requirements/base.in
    #   fastapi
pyjwt==2.4.0
    # via -r api/requirements/base.in
pystemmer==2.0.1
    # via -r api/requirements/base.in
python-dateutil==2.8.2
    # via dateparser
python-jose[cryptography]==3.3.0
    # via -r api/requirements/base.in
pytz==2022.1
    # via dateparser
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
regex==2022.3.2
    # via dateparser
requests==2.28.1
    # via cent
rfc3986[idna2008]==1.5.0
    # via httpx
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.39
    # via -r api/requirements/base.in
starlette==0.19.1
    # via fastapi
typing-extensions==4.3.0
    # via
    #   pydantic
    #   starlette
tzdata==2022.1
    # via pytz-deprecation-shim
tzlocal==4.2
    # via dateparser
urllib3==1.26.11
    # via requests
uvicorn==0.18.2
    # via -r api/requirements/base.in
zipp==3.8.1
    # via importlib-metadata
